Refactoring

+++ src/pax/core/types.hpp
@@ -41,12 +41,12 @@ namespace pax {
     using f32 = float;
 
     using u64 = unsigned long long;
-    using u32 = unsigned long;
+    using u32 = unsigned int;
     using u16 = unsigned short;
     using u8  = unsigned char;
 
     using i64 = long long;
-    using i32 = long;
+    using i32 = int;
     using i16 = short;
     using i8  = char;
